+++ .lintstagedrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  '**/*.{js,ts}': (filenames) => {
+    return [
+      `eslint --fix ${filenames.join(' ')}`,
+    ];
+  },
+};