+++ src/to-ts/index.js
@@ -1,8 +1,8 @@
-const dtsgenerator = require('dtsgenerator')
+const dtsgenerator = require('dtsgenerator');
 
 module.exports = async ({
   swagger,
-  options
+  options,
 }) => {
   const rawOutput = await dtsgenerator.default({
     contents: [swagger],
@@ -10,11 +10,11 @@ module.exports = async ({
   const lines = rawOutput.split('\n');
   return lines
     // delete unused declaration
-    .slice(1, lines.findIndex(line => line.includes('declare namespace Paths')) - 1)
+    .slice(1, lines.findIndex((line) => line.includes('declare namespace Paths')) - 1)
     // index fix
-    .map(line => line.slice(4)) 
+    .map((line) => line.slice(4))
     .join('\n')
     .replace(/\n+/g, '\n')
     // indet fix
     .replace(/ {4}/g, '  ');
-}
+};